refactor: lazy load image 위치 변경 및 id className 받아오도록 설정
