+++ client/src/Pages/Signup/index.tsx
@@ -124,7 +124,7 @@ const SignupPage = () => {
             input={pw}
             validation={pwValidation}
             placeholder="비밀번호"
-            type="password"
+            type="new-password"
             message={VALIDATION_ERR_MSG.INVALID_PW}
           />
         </InputSection>
@@ -133,7 +133,7 @@ const SignupPage = () => {
             input={pwConfirm}
             validation={confirmValidation}
             placeholder="비밀번호 확인"
-            type="password"
+            type="new-password"
             message={VALIDATION_ERR_MSG.INVALID_CONFIRM}
           />
         </InputSection>
